Treatment of Visitor Missing History
Reproduced analysis perfromed by winners on the missing values of visitor history. Bar plots show the trend 'small difference in rating/price implies higher rating probability'. Idea is to introduce dummy variables for hotels with prices/star ratings close to historical price/rating of user (of course, provided that the history is available!). Hence, the dummy is non-zero iff there is history information and a given price/rating is close to the historical values. We should expect our method, say linear regression, to set a positive coefficient on the dummy (if dummy equal 1, then hotel has higher change to be chosen).