Modifications to PredictiveAnalysis and first predictions file
The PredictiveAnalysis file has been modified in a superficial way. The code has been slighlty cleaned up and made more clear. A few variables and functions have been added to aid with the predictions of the test set, and with writing this is a new csv file. The model which generated the first predictions file was found to have NDCG of 0.462, which is lower than the current best but this had to be lowered due to the unavailability of certain features in the test set. The first predictions file is correctly formatted and is ready for submission.