Adding position features
Extra features have been generated in PredictiveAnalysis for the mean, upper and lower quartiles of each prop_id position. These are then standardised over the whole dataset and NaN values also carefully taken care of by assuming that the small number of prop_ids, around 14000, which are available in the test set and not training set, have the lowest possible value for these extra features.